Move config and dependencies
config.ts is now inside of src/ so tsc transpiles the project without a src/ folder inside of dist/.

Also, I moved all the typescript devdependencies to regular dependencies because i dont actually know how to run this project after ive transpiled it
