+++ package.json
@@ -9,24 +9,21 @@
   "main": "src/app.ts",
   "scripts": {
     "start": "ts-node src/app.ts",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "ts-node-dev --respawn --transpile-only src/app.ts"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "garlicOS®",
   "license": "ISC",
   "dependencies": {
+    "@types/node": "^14.11.7",
     "axios": "^0.20.0",
     "chalk": "^4.1.0",
     "discord.js": "^12.3.1",
     "discord.js-commando": "github:discordjs/Commando",
     "dotenv": "^8.2.0",
     "form-data": "^3.0.0",
-    "purpl-markov-chain": "^2.0.3"
-  },
-  "devDependencies": {
-    "@types/node": "^14.11.2",
-    "@types/ws": "^7.2.7",
-    "ts-node-dev": "^1.0.0-pre.63",
+    "purpl-markov-chain": "^2.0.3",
+    "ts-node": "^9.0.0",
     "typescript": "^4.0.3"
-  }
+  },
+  "devDependencies": {}
 }
