+++ src/config.template.ts
